FIR: retain substitutor for substitution override
The substitutor is needed by the IDE after commit
fa8bb47bdf0c4cea51eb63bc5a89d3e06ce6f36d. That is, `KtSymbol` now only
represents substitution override that is created by inheriting classes
that with type parameters. Use site substitution will not be captured by
a KtSymbol. As a result, the analysis API needs to use this bridging
substitutor to retain the typing information for use site substitution.

As part of this change, I also refactored
`createNewTypeParametersAndSubstitutor` to contain a boolean that
explicitly states if there are changes in type parameters. In addition,
the type parameters are set to new type parameters in all cases. This
way, in the pure FIR code path (a.k.a no Fir2Ir),
`configureAnnotationsTypeParametersAndSignature` is always invoked
with non-null type parameters. This has two effects
1. the substitutor created by `createNewTypeParametersAndSubstitutor`
   would be suitable for bridging the original member and the created
   fake override.
2. we no longer throw away type parameters when there are no changes
   and then re-create these type parameters again in
   `configureAnnotationsTypeParametersAndSignature`