FIR: track type alias on constructor symbol
Consider the following

```
class Foo

typealias Bar = Foo

fun test() {
  Bar()
}
```

The resolved symbol in `Bar()` is the constructor of `Foo`. In this case
the IDE also requires a reference to type alias `Bar` in order to make
rename refactoring work. Hence, an attribute is added to FirConstructor
to track this information.