Copy kotlin-jps-plugin-classpath artifact into kotlin-jps-plugin
This is first step towards renaming `kotlin-jps-plugin-classpath` to
`kotlin-jps-plugin`. In some places, IDE depends on old artifact so I
want to keep it for a while until I cleanup IDE completely.

This commit is going to be cherry-picked to a bunch of old Kotlin
versions, so it needs to be conservative. That's why a lot of unneeded
dependencies are packed into `kotlin-jps-plugin`. I will cleanup those
dependencies in the next commit (but the cleanup will be applied only
since 1.7.20)

(cherry picked from commit 5ee87d22305a14ce0137f477dbaa9588bd2855a5)