Refactor: Inline kotlinReflectModule
In the next commit, source dependency on kotlin-reflect will be replaced
with binary dependency. Thus, it won't be technically be a a "module"
anynmore. So I either have to rename "kotlinReflectModule" or get rid of
it. I decided to get rid of it (inline it) because there is only one
usage.