Refactor: Inline kotlinReflectModule
This is a preparation commit for the next commit. In the next commit,
source dependency on kotlin-reflect will be replaced with binary
dependency on kotlin-reflect and for some reason I can't call
`commonDependency` extension function in `kotlin-native/build.gradle`
but I can call it in `kotlin-native/backend.native/build.gradle`.